Move tests to correctly named file
